Make the serialization of test cell on the long long int type instead of long.
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/scalfmm/scalfmm/trunk@268 2616d619-271b-44dc-8df4-d4a8f33a7222

+++ Tests/testFmmAlgorithmProcPeriodic.cpp
@@ -82,7 +82,7 @@ public:
         this->dataUp = *(long long int*)buffer;
     }
 
-    static const int SerializedSizeDown = sizeof(long);
+    static const int SerializedSizeDown = sizeof(long long int);
     void serializeDown(void* const buffer) const {
         *(long long int*)buffer = this->dataDown;
     }
